+++ src/routes/Game.js
@@ -246,6 +246,7 @@ module.exports = app => {
      * @apiParam (Request Body Parameters) {Boolean} [options.roles.wildChild.isTransformationRevealed=false] If set to `true`, when `wild-child` joins the `werewolves` side because his model died, the game master will announce his transformation to other players. Default is `false`.
      * @apiParam (Request Body Parameters) {Object} [options.roles.dogWolf] Game dog wolf's role options.
      * @apiParam (Request Body Parameters) {Boolean} [options.roles.dogWolf.isChosenSideRevealed=false] If set to `true`, when `dog-wolf` chooses his side at the beginning of the game, the game master will announce the chosen side to other players. Default is `false`.
+     * @apiParam (Request Body Parameters) {Boolean} [options.roles.dogWolf.isChosenSideRandom=false] If set to `true`, the chosen side for the `dog-wolf` will be randomly defined by the API. Default is `false`.
      * @apiParam (Request Body Parameters) {Object} [options.roles.thief] Game thief's role options.
      * @apiParam (Request Body Parameters) {Boolean} [options.roles.thief.mustChooseBetweenWerewolves=true] If set to `true`, if all `thief` additional cards are from the `werewolves` side, he can't skip and must choose one. Default is `true`.
      * @apiParam (Request Body Parameters) {Number{>= 1 && <= 5}} [options.roles.thief.additionalCardsCount =2] Number of additional cards for the `thief` at the beginning of the game. Default is `2`.
@@ -366,6 +367,10 @@ module.exports = app => {
             .isBoolean().withMessage("Must be a valid boolean")
             .toBoolean(),
         body("options.roles.dogWolf.isChosenSideRevealed")
+            .optional()
+            .isBoolean().withMessage("Must be a valid boolean")
+            .toBoolean(),
+        body("options.roles.dogWolf.isChosenSideRandom")
             .optional()
             .isBoolean().withMessage("Must be a valid boolean")
             .toBoolean(),